requirement.txt
opencv-python
mediapipe
pygame
flask
flask-socketio
eventlet
requests
numpy
python-dotenv
